PlayerRemoteServer Split - builds full game
