--- src/lib-tempo/include/tempo/entity/SystemPlayerRemoteServer.hpp
@@ -1,36 +1,0 @@
-#ifndef TEMPO_SYSTEM_PLAYERREMOTESERVER_HPP
-#define TEMPO_SYSTEM_PLAYERREMOTESERVER_HPP
-
-#include <anax/System.hpp>
-#include <anax/Component.hpp>
-
-#include <tempo/component/ComponentGridMotion.hpp>
-#include <tempo/component/ComponentPlayerRemoteServer.hpp>
-#include <tempo/system/SystemID.hpp>
-#include <tempo/entity/SystemQID.hpp>
-#include <tempo/network/queue.hpp>
-#include <tempo/network/server.hpp>
-#include <tempo/time.hpp>
-
-#include <SDL.h>
-#include <iostream>
-#include <stdexcept>
-#include <cstdio>
-
-#undef main // SDL likes to define main
-
-namespace tempo{
-
-	struct SystemPlayerRemoteServer : anax::System<anax::Requires<ComponentGridMotion, 
-	                                                        ComponentPlayerRemoteServer>> 
-	{
-		tempo::Clock& clock;
-
-		inline SystemPlayerRemoteServer(tempo::Clock& c) : clock(c) {}
-
-		void advanceBeat();
-		bool update(SystemID system_id);
-	};
-}
-
-#endif
