--- src/lib-tempo/src/entity/ComponentPlayerRemoteServer.cpp
@@ -1,7 +1,0 @@
-#include <tempo/component/ComponentPlayerRemoteServer.hpp>
-
-namespace tempo{
-	
-	//Nothing Implemented
-
-}
